## What type of PR is this? (check all applicable)
- [x] Optimization

## Description
Comments and Display Ad Events are medium size tables that need to have integers updated to bigints. Each table takes between 5-10 seconds to update. I have increased the STATEMENT_TIMEOUT on production to 15 seconds to help deal with these minor delays.
[Here is the timing data against our production fork](https://gist.github.com/mstruve/1600b29a30b39b9f4edc0c8d79d9ba91)
```
# 7 seconds
"comments"
# -- change_column(:comments, :id, :bigint)
#    -> 7.1299s
# => []
# irb(main):217:1* Benchmark.realtime {
# irb(main):218:2*   ActiveRecord::Base.connection.execute(
# irb(main):219:0"     <<-SQL
# irb(main):220:0"     ALTER TABLE comments
# irb(main):221:0"     ALTER COLUMN commentable_id TYPE bigint,
# irb(main):222:0"     ALTER COLUMN user_id TYPE bigint
# irb(main):223:2*     SQL
# irb(main):224:1*     )
# irb(main):225:0> }
# => 7.011359310010448

# 6 seconds
"display_ad_events"
# irb(main):235:1* ActiveRecord::Migration[6.0].safety_assured {
# irb(main):236:1*   ActiveRecord::Migration[6.0].change_column :display_ad_events, :display_ad_id, :bigint
# irb(main):237:0> }
# -- change_column(:display_ad_events, :display_ad_id, :bigint)
#    -> 6.0153s
# => []
# irb(main):238:1* ActiveRecord::Migration[6.0].safety_assured {
# irb(main):238:1* ActiveRecord::Migration[6.0].safety_assured {
# irb(main):239:1*   ActiveRecord::Migration[6.0].change_column :display_ad_events, :user_id, :bigint
# irb(main):240:0> }
# -- change_column(:display_ad_events, :user_id, :bigint)
#    -> 5.4779s
```

`github_id_code` for GithubRepos should be a non-issue, that table is small and updates in under a second.

## Related Tickets & Documents
https://github.com/forem/forem/projects/9#card-37704279


![alt_text](https://media1.tenor.com/images/c487dd345caf07207312fcdb1a460a41/tenor.gif?itemid=13966265)